requirements.txt
streamlit
pydantic
ruamel.yaml
opencv-python
imutils
easyocr==1.7.1
torch
torchvision
